db/requirements.txt
pandas>=1.5.0
psycopg2-binary>=2.9.0
tqdm>=4.65.0
python-dotenv>=1.0.0 